requirements-devel.txt
autopep8==1.5.1
bandit==1.6.2
coverage==5.1
django-debug-toolbar==2.2
django-debug-toolbar-request-history==0.1.1
django-webtest==1.9.7
factory-boy==2.12.0
isort==4.3.21
flake8==3.7.9
flake8-bugbear==20.1.4
mock==4.0.2
pylama==7.7.1
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-env==0.6.2
pytest-mock==3.0.0
pytest-pythonpath==0.7.3
pytest-sugar==0.9.2
safety==1.8.7
